Remove semi-duplicate DWP services
Remove the process_dwp_service and test and remove
the (currently hidden) call to ProcessDwpService

As there is now a single service, there is no need for
shared components.  Refactored all code, back into single
remaining service.
